Moved GUID and UUID to metal-formats.
